Modified version for purpose of my thesis

+++ 06_dft_test.py
@@ -19,7 +19,8 @@ def test(input, n):
 
     fs = np.fft.fft(ts_np)  # Compute DFT
 
-    mags = abs(fs)[:n/2]  #Compute magnitudes of first half of sequence
+
+    mags = abs(fs)[:int(n/2)]  #Compute magnitudes of first half of sequence
 
 
     N1 = 0.0   # Count the peaks above the upper theshold
